requirements.txt
alembic==1.13.1
alphalens-reloaded==0.4.3
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
bcolz-zipline==1.2.6
beautifulsoup4==4.12.3
black==24.4.0
blosc2==2.6.2
Bottleneck==1.3.8
certifi==2024.2.2
charset-normalizer==3.3.2
clarabel==0.7.1
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cvxpy==1.4.3
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
ecos==2.0.5
empyrical-reloaded==0.5.9
et-xmlfile==1.1.0
exchange_calendars==4.5.3
executing==2.0.1
fonttools==4.51.0
frozendict==2.4.2
h5py==3.11.0
html5lib==1.1
idna==3.7
intervaltree==3.1.0
ipykernel==6.29.4
ipython==8.23.0
iso3166==2.1.1
iso4217==1.11.20220401
jedi==0.19.1
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lru-dict==1.3.0
lxml==5.2.1
Mako==1.3.3
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
msgpack==1.0.8
multipledispatch==1.0.0
multitasking==0.0.11
mypy==1.9.0
mypy-extensions==1.0.0
ndindex==1.8
nest-asyncio==1.6.0
networkx==3.3
numexpr==2.10.0
numpy==1.26.4
openpyxl==3.1.2
osqp==0.6.5
packaging==24.0
pandas==2.2.2
pandas-datareader==0.10.0
pandas-stubs==2.2.1.240316
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
peewee==3.17.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pybind11==2.12.0
pyfolio-reloaded==0.9.5
Pygments==2.17.2
pyluach==2.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-interface==1.6.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
qdldl==0.1.7.post2
requests==2.31.0
scikit-learn==1.4.2
scipy==1.11.4
scs==3.2.4.post1
seaborn==0.13.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
statsmodels==0.14.2
TA-Lib==0.4.28
tables==3.9.2
threadpoolctl==3.4.0
toolz==0.12.1
tornado==6.4
trading_calendars==2.1.1
traitlets==5.14.3
types-pytz==2024.1.0.20240417
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yfinance==0.2.38
zipline-reloaded==3.0.3
